+++ api/opentrons/broker/broker.py
@@ -1,6 +1,4 @@
 import asyncio
-import functools
-import inspect
 
 from asyncio import Queue
 from concurrent import futures
@@ -48,45 +46,6 @@ class Notifications(object):
         return self
 
 
-def publish(before, after, name, **decorator_kwargs):
-    def decorator(f):
-        @functools.wraps(f)
-        def decorated(*args, **kwargs):
-            _args = _get_args(f, args, kwargs)
-            _args.update(decorator_kwargs)
-
-            if before:
-                notify(name, {**_args, '$': 'before'})
-
-            res = f(*args, **kwargs)
-
-            if after:
-                notify(name, {**_args, '$': 'after', 'return': res})
-
-            return res
-        return decorated
-
-    return decorator
-
-
-def _get_args(f, args, kwargs):
-    # Create the initial dictionary with args that have defaults
-    res = {}
-
-    if inspect.getargspec(f).defaults:
-        res = dict(
-            zip(
-                reversed(inspect.getargspec(f).args),
-                reversed(inspect.getargspec(f).defaults)))
-
-    # Update / insert values for positional args
-    res.update(dict(zip(inspect.getargspec(f).args, args)))
-
-    # Update it with values for named args
-    res.update(kwargs)
-    return res
-
-
 def subscribe(topics, handler=None, loop=None):
     notifications = None
 
@@ -109,8 +68,3 @@ def notify(name, payload):
     for topics, listener in listeners:
         if name in topics:
             listener(name, payload)
-
-
-before = functools.partial(publish, before=True, after=False)
-after = functools.partial(publish, before=False, after=True)
-both = functools.partial(publish, before=True, after=True)
