+++ api/opentrons/instruments/pipette.py
@@ -8,7 +8,7 @@ from opentrons.containers.placeable import (
 )
 from opentrons.helpers import helpers
 from opentrons.instruments.instrument import Instrument
-from opentrons.broker import both
+from opentrons.broker import publish
 
 
 class Pipette(Instrument):
@@ -274,7 +274,7 @@ class Pipette(Instrument):
 
         return self
 
-    @both(
+    @publish.both(
         name='robot.command',
         text='Aspirating {volume} uL from {location} at {rate} speed'
     )
@@ -360,7 +360,7 @@ class Pipette(Instrument):
         self.current_volume += volume  # update after actual aspirate
         return self
 
-    @both(
+    @publish.both(
         name='robot.command', text='Dispensing {volume} into {location}'
     )
     def dispense(self,
@@ -466,7 +466,7 @@ class Pipette(Instrument):
                 location = location.bottom(min(location.z_size(), 1))
             self.move_to(location, strategy='direct')
 
-    @both(
+    @publish.both(
         name='robot.command',
         text='Mixing {repetitions} times with a volume of {volume}ul')
     def mix(self,
@@ -534,7 +534,7 @@ class Pipette(Instrument):
 
         return self
 
-    @both(name='robot.command', text='Blowing out at {location}')
+    @publish.both(name='robot.command', text='Blowing out at {location}')
     def blow_out(self, location=None):
         """
         Force any remaining liquid to dispense, by moving
@@ -571,7 +571,7 @@ class Pipette(Instrument):
 
         return self
 
-    @both(name='robot.command', text='Touching tip')
+    @publish.both(name='robot.command', text='Touching tip')
     def touch_tip(self, location=None, radius=1.0):
         """
         Touch the :any:`Pipette` tip to the sides of a well,
@@ -639,7 +639,7 @@ class Pipette(Instrument):
 
         return self
 
-    @both(name='robot.command', text='Air gap')
+    @publish.both(name='robot.command', text='Air gap')
     def air_gap(self, volume=None, height=None):
         """
         Pull air into the :any:`Pipette` current tip
@@ -689,7 +689,7 @@ class Pipette(Instrument):
         self.aspirate(volume)
         return self
 
-    @both(name='robot.command', text='Returning tip')
+    @publish.both(name='robot.command', text='Returning tip')
     def return_tip(self, home_after=True):
         """
         Drop the pipette's current tip to it's originating tip rack
@@ -729,7 +729,7 @@ class Pipette(Instrument):
         self.drop_tip(self.current_tip(), home_after=home_after)
         return self
 
-    @both(name='robot.command', text='Picking up tip {location}')
+    @publish.both(name='robot.command', text='Picking up tip {location}')
     def pick_up_tip(self, location=None, presses=3):
         """
         Pick up a tip for the Pipette to run liquid-handling commands with
@@ -795,7 +795,7 @@ class Pipette(Instrument):
 
         return self
 
-    @both(name='robot.command', text='Dropping tip {location}')
+    @publish.both(name='robot.command', text='Dropping tip {location}')
     def drop_tip(self, location=None, home_after=True):
         """
         Drop the pipette's current tip
@@ -880,7 +880,7 @@ class Pipette(Instrument):
         >>> p200.home() # doctest: +ELLIPSIS
         <opentrons.instruments.pipette.Pipette object at ...>
         """
-        @both(
+        @publish.both(
             name='robot.command',
             text='Homing pipette plunger on axis {axis}')
         def _home(axis):
@@ -890,7 +890,7 @@ class Pipette(Instrument):
         _home(self.axis)
         return self
 
-    @both(
+    @publish.both(
         name='robot.command',
         text='Distributing {volume} from {source} to {dest}')
     def distribute(self, volume, source, dest, *args, **kwargs):
@@ -921,7 +921,7 @@ class Pipette(Instrument):
             kwargs['disposal_vol'] = self.min_volume
         return self.transfer(*args, **kwargs)
 
-    @both(
+    @publish.both(
         name='robot.command',
         text='Consolidating {volume} from {source} to {dest}')
     def consolidate(self, volume, source, dest, *args, **kwargs):
@@ -952,7 +952,7 @@ class Pipette(Instrument):
         args = [volume, source, dest, *args]
         return self.transfer(*args, **kwargs)
 
-    @both(
+    @publish.both(
         name='robot.command',
         text='Transferring {volume} from {source} to {dest}')
     def transfer(self, volume, source, dest, **kwargs):
@@ -1070,7 +1070,7 @@ class Pipette(Instrument):
 
         return self
 
-    @both(
+    @publish.both(
         name='robot.command',
         text='Delaying for {minutes}m {seconds}s')
     def delay(self, seconds=0, minutes=0):
